Rename StorageManager to Storage
