+++ src/main/java/mryapper/command/Command.java
@@ -1,6 +1,6 @@
 package mryapper.command;
 
-import mryapper.storagemanager.StorageManager;
+import mryapper.storage.Storage;
 import mryapper.task.TaskList;
 
 public abstract class Command {
@@ -9,8 +9,8 @@ public abstract class Command {
      * Executes the command.
      *
      * @param tasks The TaskList to execute the command on.
-     * @param storageManager The Storage to execute the command on.
+     * @param storage The Storage to execute the command on.
      * @return The response from the ChatBot to the user.
      */
-    public abstract String execute(TaskList tasks, StorageManager storageManager);
+    public abstract String execute(TaskList tasks, Storage storage);
 }
